rename api_client to service_api_client
