+++ tests/app/notify_client/test_service_api_client.py
@@ -1,4 +1,4 @@
-from app.notify_client.api_client import ServiceAPIClient
+from app.notify_client.service_api_client import ServiceAPIClient
 from tests.conftest import fake_uuid
 
 
@@ -13,8 +13,8 @@ def test_client_posts_archived_true_when_deleting_template(mocker):
     expected_url = '/service/{}/template/{}'.format(service_id, template_id)
 
     client = ServiceAPIClient()
-    mock_post = mocker.patch('app.notify_client.api_client.ServiceAPIClient.post')
-    mock_attach_user = mocker.patch('app.notify_client.api_client._attach_current_user',
+    mock_post = mocker.patch('app.notify_client.service_api_client.ServiceAPIClient.post')
+    mock_attach_user = mocker.patch('app.notify_client.service_api_client._attach_current_user',
                                     side_effect=lambda x: x.update({'created_by': fake_uuid}))
 
     client.delete_service_template(service_id, template_id)
